pyproject.toml
[project]
name = "vibe-trade-agent"
version = "0.1.0"
description = "AI agent for creating trading strategies using Vibe Trade MCP"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.0.14",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]

[project.scripts]
main = "src.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

